src/requirements.txt
web.py
feedparser
requests
google-cloud-storage
google-cloud-aiplatform